python/llms/anthropic-structured-outputs/requirements.txt
anthropic==0.34.1
flask==3.0.3
instructor==1.4.0
openlayer>=0.2.0a21
python-dotenv==1.0.1
